Major Multiplayer commit
This commit adds major things.
Host-Client communication for up to 16 Players.
Adds the ability for all enemies to appear a uniform, different color
from yourself in your view window.
adds the ability to shoot a basic bullet, with health bars
Changed the Camera angle to 60 degrees from the horizon down so that it
is an angled top-down shooter.